feat(lib): Add support for RTL diff images
If the base and change image for a diff render have different dimensions
they were previously both aligned to the top left. This allows configureing the
common origin to be another of the four corners.

The cli now accepts a `--dir={ltr,rtl}` option to chose between top left and top
right to improve mismatched diff images for rtl documents.